git hey error

+++ app.js
@@ -68,14 +68,13 @@ document.getElementById('book-form').addEventListener('submit', (e) => {
     const book = new Book (title, author, isbn);
     console.log(book);
 
-    const ui = new UI();
-
     if(title === '' || author === '' || isbn ==='') {
         ui.showAlert('Niste popunili sva polja', 'error');
     } else if (ui.isBookExists(isbn)) {
         ui.showAlert('Knjiga sa ovim ISBN-om vec postoji!', 'error');
     } else {
         ui.addBookToList(book);
+        Storage.addBook(book) // Dodavanje knjige u storage
         ui.showAlert('Uspesno ste dodali knjigu!', 'success');
         ui.clearFields();
     }
@@ -84,15 +83,41 @@ document.getElementById('book-form').addEventListener('submit', (e) => {
 // Listener za delete
 document.getElementById('book-list').addEventListener('click', e => {
     e.preventDefault();
-    const ui = new UI();
-    ui.showAlert('Uspesno ste obrisali knjigu!', 'error')
-    ui.deleteBook(e.target)
+    if (e.target.className === 'delete') {
+        ui.deleteBook(e.target);
+        Storage.removeBook(e.target.parentElement.previousElementSibling.textContent);        
+        ui.showAlert('Uspesno ste obrisali knjigu!', 'success')
+    }
+    
 })
+// Local Storage
+class Storage {
+    static getBooks() {
+        let books = localStorage.getItem('books');
+        return books ? JSON.parse(books) : [];
+    };
 
+    static addBook(book) {
+        const books = Storage.getBooks();
+        books.push(book);
+        localStorage.setItem('books', JSON.stringify(books));
+    };
 
+    static removeBook(isbn) {
+        const books = Storage.getBooks();
+        const filteredBooks = books.filter(book => book.isbn !== isbn);
+        localStorage.setItem('books', JSON.stringify(filteredBooks));
+    }
+}
+// Globalni UI(optimizacija)
+const ui = new UI();
 
+// eListener za storage
+document.addEventListener('DOMContentLoaded', () => {
+    const books = Storage.getBooks();
+    books.forEach(book => ui.addBookToList(new Book(book.title, book.author, book.isbn)));
+});
 
-// Neki local storage da sacuva knjige.
 
 /*  
     - Poboljšanje korisničkog interfejsa:
